Fix: Corrección de errores y restricciones

+++ cocemfe_sevilla/professionals/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.urls import reverse_lazy
 from django.views import View
 from django.views.generic import DeleteView, DetailView, UpdateView
+from django.contrib import messages
 
 from .models import Professional
 from .forms import ProfessionalForm
@@ -18,8 +19,11 @@ class EditUserView(View):
         print(request.FILES)
         form = ProfessionalForm(request.POST,request.FILES, instance=professional)
         if form.is_valid():
-            form.save()
-            return redirect('/professionals/?message=Profesional editado&status=Success')
+            if request.user.is_superuser:
+                form.save()
+                return redirect('/professionals/?message=Profesional editado&status=Success')
+            else:
+                return render(request, '403.html')
         else:
             return render(request, 'professional_detail.html', {'form': form, 'professional': professional})
 
@@ -53,12 +57,17 @@ def professional_list(request):
 
 def delete_professional(request, id):
     professional = get_object_or_404(Professional, id=id)
-    print(professional)
+    professionals = Professional.objects.all()
 
     if request.method == 'POST':
-        professional.delete()
-
-    professionals = Professional.objects.all()
+        if request.user.is_superuser:
+            professional.delete()
+            messages.success(request, 'Profesional eliminado exitosamente.')\
+            
+            return render(request, 'professional_list.html', {'professionals': professionals})
+        else:
+            return render(request, '403.html')
+         
 
     return render(request, 'professional_list.html', {'professionals': professionals})
